Fixes handling of nkBracketExpr nodes in opcGetImpl.
This fixes a problem that is similar to #8484. I attempted to
reproduce it via:

```nim
import macros

template custom {.pragma.}

type
  Vec[T] = object
    x, y: T
  Entry = object
    c: Vec[float]
    a {.custom.}: int
    b: float

proc foo(o: object) =
  for name, value in o.fieldPairs:
    when value.hasCustomPragma(custom):
      echo(name, " has {.custom.}")

var x: Entry
foo(x)
```

But was not able to do so, possibly just a case of my code having
nesting of templates/generics. But this simple patch fixes the
compilation of my code.